Update README.md
NWOCOINPROJEC WAS development tree

NWOCOIN (Nwocoin.org) was a PoS-based cryptocurrency. has now been migrated to Binance Smart Chain(BSC BEP-20).

Development process
===========================

Nwocoin (NWO) is a meme cryptocurrency, deployed on Binance Smart Chain(BSC BEP-20).
Nwocoin is an experimental digital currency thatenables instant payments to anyone, anywhere
in the world.
Nwocoin is a meme and secureand anonymous cryptocurrency, built with a focus on privacy.
We are Planning to create an environment that accepts any
credit or debit card from VISA or MasterCard. This option may be very useful and Creates a
community and projects that attract more investment. Nwocoin is a digital virtual
currency that uses cryptography for security.